+++ src/stories/navigation/header/_types.tsx
@@ -1,0 +1,6 @@
+import { IHeaderProps } from '@zendeskgarden/react-chrome';
+
+export interface HeaderArgs extends IHeaderProps {
+    /** Displays logo for standlone usage  */
+    isStandalone?: boolean;
+}